Fix sources root and clean up duplicate src folder
