--- .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Wed Jan 15 09:45:35 KST 2025
-gradle.version=8.11.1
