--- .idea/easycode.ignore
@@ -1,13 +1,0 @@
-.idea
-.vscode
-node_modules/
-dist/
-vendor/
-cache/
-.*/
-*.min.*
-*.test.*
-*.spec.*
-*.bundle.*
-*.bundle-min.*
-*.log
