--- .idea/modules.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectModuleManager">
-    <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/NBE3-4-1-Team15.iml" filepath="$PROJECT_DIR$/.idea/modules/NBE3-4-1-Team15.iml" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/NBE3-4-1-Team15.main.iml" filepath="$PROJECT_DIR$/.idea/modules/NBE3-4-1-Team15.main.iml" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/NBE3-4-1-Team15.test.iml" filepath="$PROJECT_DIR$/.idea/modules/NBE3-4-1-Team15.test.iml" />
-    </modules>
-  </component>
-</project>
