--- README.md
@@ -1,2 +1,0 @@
-# NBE3-4-1-Team15
-프로그래머스 백엔드 데브코스 3기 4회차 15팀 1차 프로젝트 repository입니다.
