+++ backend/src/main/java/com/nbe/NBE3_4_1_Team15/domain/cart/repository/CartRepository.java
@@ -1,0 +1,7 @@
+package src.main.java.com.nbe.NBE3_4_1_Team15.domain.cart.repository;
+
+import src.main.java.com.nbe.NBE3_4_1_Team15.domain.cart.entity.Cart;
+
+// CartRepository.java git 업로드용 주석 추가
+public interface CartRepository extends JpaRepository<Cart, Long> {
+}
