+++ backend/src/main/java/com/nbe/NBE3_4_1_Team15/domain/cartProduct/entity/CartProduct.java
@@ -10,6 +10,10 @@ import lombok.Getter;
 
 @Entity
 @Getter
+@Setter
+@NoArgsConstructor // JPA에서 기본 생성자 필요
+@AllArgsConstructor // 모든 필드를 사용하는 생성자 자동 생성
+// CartProduct.java git 업로드용 주석 추가
 public class CartProduct extends BaseTime {
     @ManyToOne(fetch = FetchType.LAZY)
     private Cart cart;
