+++ backend/src/main/java/com/nbe/NBE3_4_1_Team15/domain/cartProduct/repository/CartProductRepository.java
@@ -1,0 +1,7 @@
+package src.main.java.com.nbe.NBE3_4_1_Team15.domain.cartProduct.repository;
+
+import com.nbe.NBE3_4_1_Team15.domain.cartProduct.entity.CartProduct;
+
+public interface CartProductRepository extends JpaRepository<CartProduct, Long> {
+}
+
