--- settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'NBE3-4-1-Team15'
