--- src/main/java/com/nbe/NBE3_4_1_Team15/domain/member/type/MemberType.java
@@ -1,6 +1,0 @@
-package com.nbe.NBE3_4_1_Team15.domain.member.type;
-
-public enum MemberType {
-    ADMIN,
-    CUSTOMER
-}
