--- src/main/java/com/nbe/NBE3_4_1_Team15/domain/order/entity/Order.java
@@ -1,24 +1,0 @@
-package com.nbe.NBE3_4_1_Team15.domain.order.entity;
-
-import com.nbe.NBE3_4_1_Team15.domain.cart.entity.Cart;
-import com.nbe.NBE3_4_1_Team15.domain.member.entity.Member;
-import com.nbe.NBE3_4_1_Team15.global.jpa.entity.BaseTime;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(name="order_table")
-public class Order extends BaseTime {
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Member consumer;
-
-    private Boolean orderStatus;
-
-    private Integer totalPrice;
-}
