--- src/main/java/com/nbe/NBE3_4_1_Team15/global/jpa/entity/BaseEntity.java
@@ -1,21 +1,0 @@
-package com.nbe.NBE3_4_1_Team15.global.jpa.entity;
-
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.MappedSuperclass;
-import lombok.AccessLevel;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@EqualsAndHashCode(onlyExplicitlyIncluded = true)
-@MappedSuperclass
-public class BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Setter(AccessLevel.PRIVATE)
-    @EqualsAndHashCode.Include
-    private Long id;
-}
