--- src/main/java/com/nbe/NBE3_4_1_Team15/global/jpa/entity/BaseTime.java
@@ -1,25 +1,0 @@
-package com.nbe.NBE3_4_1_Team15.global.jpa.entity;
-
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@Getter
-@EntityListeners(AuditingEntityListener.class)
-@MappedSuperclass
-public class BaseTime extends BaseEntity {
-    @CreatedDate
-    @Setter(AccessLevel.PRIVATE)
-    private LocalDateTime createAt;
-
-    @LastModifiedDate
-    @Setter(AccessLevel.PRIVATE)
-    private LocalDateTime updateAt;
-}
