--- src/main/resources/application-test.yml
@@ -1,3 +1,0 @@
-spring:
-  datasource:
-    url: jdbc:h2:mem:db_test;MODE=MySQL
