--- src/main/resources/static/index.html
@@ -1,1 +1,0 @@
-test
