--- src/test/java/com/nbe/NBE3_4_1_Team15/Nbe341Team15ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.nbe.NBE3_4_1_Team15;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Nbe341Team15ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
