--- src/test/java/com/nbe/NBE3_4_1_Team15/domain/cartProduct/entity/CartProductTest.java
@@ -1,10 +1,0 @@
-package com.nbe.NBE3_4_1_Team15.domain.cartProduct.entity;
-
-import com.nbe.NBE3_4_1_Team15.domain.cart.entity.Cart;
-import com.nbe.NBE3_4_1_Team15.domain.product.entity.Product;
-import org.junit.jupiter.api.Test;
-
-public class CartProductTest {
-
-}
-
